MNES-1238: MR Suggestions (move Prisma, catches in RespositorySearchService)
