+++ src/lib/services/connection-access/PrismaConnectorInterface.ts
@@ -1,5 +1,5 @@
 import { ConnectionType } from '@prisma/client';
-import { DataSourceFormData } from 'lib/services/repository-access/PrismaConnector';
+import { DataSourceFormData } from 'lib/services/connection-access/PrismaConnector';
 
 export interface IPrismaConnector {
     getConnectionData(): unknown;
